+++ tests/testthat/test-init.R
@@ -269,3 +269,19 @@ test_that("init() respects Remotes in a project DESCRIPTION file", {
   init()
   expect_true(renv_package_installed("skeleton"))
 })
+
+test_that("a project using named remotes can be initialized", {
+  project <- renv_tests_scope()
+  
+  contents <- heredoc('
+    Depends:
+      toast
+    Remotes:
+      toast=toast
+  ')
+  writeLines(contents, con = "DESCRIPTION")
+  
+  init(settings = list(snapshot.type = "explicit"))
+  expect_true(renv_package_installed("toast"))
+  
+})